Removing the API surface from Math.CoreRT.cs and MathF.CoreRT.cs that is now shared.
